[Fix] resolve duplication in KNeighborsClassifier and KNeighborsRegressor
